refactor: 해당 패키지 admin으로 변경
